+++ Spark/lib/splog/Platform/Windows/spWindowsConfig.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "splogCore.h"
-
-namespace splog {
-
-}