+++ Spark/lib/splog/splogConfig.h
@@ -2,16 +2,39 @@
 
 #include "splogCore.h"
 
-extern enum LogType;
+#ifdef _WIN32
+	#include "Platform/Windows/splogWindowsConfig.h"
+#elif __unix__
+
+#elif __APPLE__
+
+#elif __linux__
+
+#endif
+
+#include "Format/splogPatternFormat.h"
+
+/* TODO (further fix) Enable colorizing only a segment of text in the line
+ * TODO Setting up logging format 
+ * 
+ * */
 
 namespace splog {
 
-	void configLogType(LogType type)
+	void spSetupLoggingFormat(std::string& str)
 	{
 
 	}
 
-	void configLogFormat(std::string& str)
+
+	extern void spConfigLogColor(LogType type);
+
+	void spConfigLogType(LogType type)
+	{
+		spConfigLogColor(type);
+	}
+
+	void spConfigLogFormat(std::string& str)
 	{
 		std::cout << str << std::endl;
 	}