+++ Spark/src/SparkleRel/MainEntry.h
@@ -1,7 +1,9 @@
 #pragma once
 
 #include "sppch.h"
-#include "splog.h"
+#include "lib/splog/splog.h"
+
+#include "Windows.h"
 
 extern Spark::Application* createApp();
 
@@ -9,7 +11,12 @@ int main()
 {
 	std::cout << "Welcome to Spark!" << std::endl;
 
-	SPLOG_INFO("Hello SPLOG!");
+	std::string str("123");
+	SPLOG_WARNING("Hello SPLOG!, {0}, {1}, {0}", 1, str);
+	// std::cout << splog::spToString(str) << std::endl;
+	
+	HANDLE hConsole = GetStdHandle(STD_OUTPUT_HANDLE);
+	// you can loop k higher to see more color choices
 
 	auto app = createApp();
 	app->tickUpdate();