+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 ### [22/3/22] SPLOG Basics
 * Constructed Splog framework  
 * Can now log strings properly  
-**TODO** Enable variable logging  
-**TODO** Enable setting up logging format  
-**TODO** Enable variant color for different levels of logging
+* **TODO** Enable variable logging  
+* **TODO** Enable setting up logging format  
+* **TODO** Enable variant color for different levels of logging
+
+### [22/3/23] SPLOG Improvements
+* **FIX** Variable logging
+* **FIX** Colorize message (only for whole line)